Created generic TasksTable component
